Remove coercion local inspections and tests Update coercion annotator tests to check for the local fixes (explicit/implicit)
