+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/ProofsIdePlugin.kt
@@ -3,11 +3,10 @@ package arrow.meta.ide.plugins.proofs
 import arrow.meta.ide.IdeMetaPlugin
 import arrow.meta.ide.IdePlugin
 import arrow.meta.ide.invoke
-import arrow.meta.ide.plugins.proofs.annotators.coercionAnnotator
+import arrow.meta.ide.plugins.proofs.annotators.coercion.coercionAnnotator
 import arrow.meta.ide.plugins.proofs.annotators.givenAnnotator
 import arrow.meta.ide.plugins.proofs.annotators.refinementAnnotator
 import arrow.meta.ide.plugins.proofs.folding.codeFolding
-import arrow.meta.ide.plugins.proofs.inspections.coercions.coercionInspections
 import arrow.meta.ide.plugins.proofs.markers.proofLineMarkers
 import arrow.meta.ide.plugins.proofs.markers.refinementLineMarkers
 import arrow.meta.ide.plugins.proofs.psi.isCoercionProof
@@ -35,7 +34,6 @@ val IdeMetaPlugin.typeProofsIde: IdePlugin
       addDiagnosticSuppressorWithCtx { suppressProvenTypeMismatch(it) },
       givenAnnotator,
       coercionAnnotator,
-//      coercionInspections,
       codeFolding
     )
   }
